bump to v1.0.0
At this stage, any changes to jugmt are expected to result from
modifications in the output format provided by 'senfd' or the
introduction of new input formats, such as HTML or PDF.

The prototype aims to quantify the amount of work and code required to
extract tables from .docx files and organize them into another format,
as well as to demonstrate how such a tool could be continuously tested
and deployed.

For this approach, approximately 155 lines of code are needed when
relying on third-party XML manipulation and data modeling. This makes
"gluing" these components together relatively simple, at least for .docx
files.

Thus, we are marking this with a v1 release, as the prototyping of
the extraction phase is finalized. For ease of coverage testing, the
extracting logic has been integrated into senfd. This integration allows
for more comprehensive testing of the FigureDocument construction.

It will be interesting to observe when jugmt starts to break, as the CI
currently runs daily and the dependency on senfd is unbounded, making
eventual breakage expected.

Signed-off-by: Simon A. F. Lund <os@safl.dk>